Harmonic Mean

Compute the harmonic mean

ds_hmean(x, na.rm = FALSE, ...)

Arguments

x
a numeric vector containing the values whose harmonic mean is to be computed

na.rm
a logical value indicating whether NA values should be stripped before the computation proceeds.

...
further arguments passed to or from other methods #' @details Any NA values are stripped from x before computation takes place.

Value

Returns the harmonic mean of x

See also

Examples

ds_hmean(mtcars$mpg)

#> [1] 18.44092