Multiple One & Two Way Tables

ds_oway_tables creates multiple one way tables by creating a frequency table for each categorical variable in a data frame. ds_tway_tables creates multiple two way tables by creating a cross table for each unique pair of categorical variables in a data frame.

ds_oway_tables(data)

ds_tway_tables(data)

Arguments

data

a data frame

Details

ds_oway_tables is a extension of the ds_freq_table function. It creates a frequency table for each categorical variable in the dataframe. ds_tway_tables is a extension of the ds_cross_table function. It creates a two way table for each unique pair of categorical variables in the dataframe.

Deprecated Functions

oway_tables() and tway_tables() have been deprecated. Instead use ds_oway_tables() and ds_tway_tables().

See also

link{ds_freq_table} link{ds_cross_table}

Examples

mt <- mtcars
mt[, c(2, 8, 9)] <- lapply(mt[, c(2, 8, 9)], factor)
# multiple one way tables
ds_oway_tables(mt)

#>                                Variable: cyl                                 
#> |--------------------------------------------------------------------------|
#> |                                Cumulative                    Cumulative  |
#> |    Levels    |  Frequency   |   Frequency  |   Percent    |    Percent   |
#> |--------------------------------------------------------------------------|
#> |       4      |      11      |      11      |     34.38    |     34.38    |
#> |--------------------------------------------------------------------------|
#> |       6      |       7      |      18      |     21.88    |     56.25    |
#> |--------------------------------------------------------------------------|
#> |       8      |      14      |      32      |     43.75    |      100     |
#> |--------------------------------------------------------------------------|
#> 
#> 
#>                                 Variable: vs                                 
#> |--------------------------------------------------------------------------|
#> |                                Cumulative                    Cumulative  |
#> |    Levels    |  Frequency   |   Frequency  |   Percent    |    Percent   |
#> |--------------------------------------------------------------------------|
#> |       0      |      18      |      18      |     56.25    |     56.25    |
#> |--------------------------------------------------------------------------|
#> |       1      |      14      |      32      |     43.75    |      100     |
#> |--------------------------------------------------------------------------|
#> 
#> 
#>                                 Variable: am                                 
#> |--------------------------------------------------------------------------|
#> |                                Cumulative                    Cumulative  |
#> |    Levels    |  Frequency   |   Frequency  |   Percent    |    Percent   |
#> |--------------------------------------------------------------------------|
#> |       0      |      19      |      19      |     59.38    |     59.38    |
#> |--------------------------------------------------------------------------|
#> |       1      |      13      |      32      |     40.62    |      100     |
#> |--------------------------------------------------------------------------|
#> 
#> 

# multiple two way tables
ds_tway_tables(mt)

#>     Cell Contents
#>  |---------------|
#>  |     Frequency |
#>  |       Percent |
#>  |       Row Pct |
#>  |       Col Pct |
#>  |---------------|
#> 
#>  Total Observations:  32 
#> 
#>                          cyl vs vs                           
#> -------------------------------------------------------------
#> |              |                     vs                     |
#> -------------------------------------------------------------
#> |          cyl |            0 |            1 |    Row Total |
#> -------------------------------------------------------------
#> |            4 |            1 |           10 |           11 |
#> |              |        0.031 |        0.312 |              |
#> |              |         0.09 |         0.91 |         0.34 |
#> |              |         0.06 |         0.71 |              |
#> -------------------------------------------------------------
#> |            6 |            3 |            4 |            7 |
#> |              |        0.094 |        0.125 |              |
#> |              |         0.43 |         0.57 |         0.22 |
#> |              |         0.17 |         0.29 |              |
#> -------------------------------------------------------------
#> |            8 |           14 |            0 |           14 |
#> |              |        0.438 |            0 |              |
#> |              |            1 |            0 |         0.44 |
#> |              |         0.78 |            0 |              |
#> -------------------------------------------------------------
#> | Column Total |           18 |           14 |           32 |
#> |              |        0.563 |        0.437 |              |
#> -------------------------------------------------------------
#> 
#> 
#>                          cyl vs am                           
#> -------------------------------------------------------------
#> |              |                     am                     |
#> -------------------------------------------------------------
#> |          cyl |            0 |            1 |    Row Total |
#> -------------------------------------------------------------
#> |            4 |            3 |            8 |           11 |
#> |              |        0.094 |         0.25 |              |
#> |              |         0.27 |         0.73 |         0.34 |
#> |              |         0.16 |         0.62 |              |
#> -------------------------------------------------------------
#> |            6 |            4 |            3 |            7 |
#> |              |        0.125 |        0.094 |              |
#> |              |         0.57 |         0.43 |         0.22 |
#> |              |         0.21 |         0.23 |              |
#> -------------------------------------------------------------
#> |            8 |           12 |            2 |           14 |
#> |              |        0.375 |        0.062 |              |
#> |              |         0.86 |         0.14 |         0.44 |
#> |              |         0.63 |         0.15 |              |
#> -------------------------------------------------------------
#> | Column Total |           19 |           13 |           32 |
#> |              |        0.594 |        0.406 |              |
#> -------------------------------------------------------------
#> 
#> 
#>                           vs vs am                           
#> -------------------------------------------------------------
#> |              |                     am                     |
#> -------------------------------------------------------------
#> |           vs |            0 |            1 |    Row Total |
#> -------------------------------------------------------------
#> |            0 |           12 |            6 |           18 |
#> |              |        0.375 |        0.188 |              |
#> |              |         0.67 |         0.33 |         0.56 |
#> |              |         0.63 |         0.46 |              |
#> -------------------------------------------------------------
#> |            1 |            7 |            7 |           14 |
#> |              |        0.219 |        0.219 |              |
#> |              |          0.5 |          0.5 |         0.44 |
#> |              |         0.37 |         0.54 |              |
#> -------------------------------------------------------------
#> | Column Total |           19 |           13 |           32 |
#> |              |        0.594 |        0.407 |              |
#> -------------------------------------------------------------
#> 
#>